Split tags functionality into its own module
The utilities for filtering tags are related to the loader, but can
live in their own module.

This refactor is a first step towards making the mechanism more
generic, and making the loader just use a stable interface for
performing the filtering.
